Move DisplayDataset into frontend directory
